requirements.txt
feedparser>=6.0.0
pyyaml>=6.0
requests>=2.31.0

# Development dependencies
pytest>=9.0.0
pytest-cov>=7.0.0
black>=25.0.0
mypy>=1.19.0
